[bugfix] jest のテストが終わらないバグを修正
参考: https://qiita.com/yuta-katayama-23/items/29b91fed629fe758a42d#%E2%85%B1-applisten%E3%81%AE%E9%83%A8%E5%88%86%E3%81%A7%E3%82%A8%E3%83%A9%E3%83%BC
